Trait appium_client::capabilities::AppCapable

pub trait AppCapable: AppiumCapability {
    // Provided methods
    fn app(&mut self, app_path: &str) { ... }
    fn other_apps(&mut self, paths: &[&str]) { ... }
    fn no_reset(&mut self, no_reset: bool) { ... }
    fn full_reset(&mut self, full_reset: bool) { ... }
    fn print_page_source_on_find_failure(&mut self, value: bool) { ... }
}

Capabilities for drivers that are used to run an app.

Provided Methods

fn app(&mut self, app_path: &str)

The path to an installable application.

fn other_apps(&mut self, paths: &[&str])

App or list of apps (as a JSON array) to install prior to running tests.

Note that it will not work with automationName of Espresso and iOS real devices

fn no_reset(&mut self, no_reset: bool)

Don’t reset app state before this session.

“Reset” means to delete app data (like a fresh install). If true, instruct an Appium driver to avoid its usual reset logic during session start and cleanup (default false).

fn full_reset(&mut self, full_reset: bool)

Perform a complete reset.

“Complete reset” usually means a reinstall. If true, instruct an Appium driver to augment its usual reset logic with additional steps to ensure maximum environmental reproducibility (default false)

fn print_page_source_on_find_failure(&mut self, value: bool)

When a find operation fails, print the current page source. Defaults to false.

When the element you’re looking for is not found on screen, then this setting will print DOM of the visible app screen. This DOM can be further inspected to check if the locator is correct, or if the correct page is displayed.