Trait appium_client::capabilities::AppiumCapability

pub trait AppiumCapabilitywhere
    Self: Deref<Target = Capabilities> + DerefMut<Target = Capabilities>,{
    // Provided methods
    fn automation_name(&mut self, automation_name: &str) { ... }
    fn platform_version(&mut self, version: &str) { ... }
    fn device_name(&mut self, device_name: &str) { ... }
    fn set_str(&mut self, name: &str, value: &str) { ... }
    fn set_number(&mut self, name: &str, value: Number) { ... }
    fn set_bool(&mut self, name: &str, value: bool) { ... }
}

Extensions to easily define capabilities for Appium driver. See https://appium.io/docs/en/2.1/guides/caps/.

Provided Methods

fn automation_name(&mut self, automation_name: &str)

Set the automation driver to use (the engine for tests, eg. XCuiTest for iOS).

Appium usually autoselects the driver based on platform, but you choose the preferred driver. See automation for possible values.

fn platform_version(&mut self, version: &str)

The version of a platform, e.g., for iOS, “16.0”

fn device_name(&mut self, device_name: &str)

The name of a particular device to automate.

For example “iPhone 14”. Currently only actually useful for specifying iOS simulators, since in other situations it’s typically recommended to use a specific device id via the appium:udid capability.

fn set_str(&mut self, name: &str, value: &str)

Sets a string capability.

For example set_str("appium:deviceName", "iPhone 14").

fn set_number(&mut self, name: &str, value: Number)

Sets a number capability.

For example set_number("appium:newCommandTimeout", Number::from(120u64)).

fn set_bool(&mut self, name: &str, value: bool)

Sets a boolean capability.

For example set_bool("appium:noReset", true)