Trait appium_client::capabilities::XCUITestAppCompatible

pub trait XCUITestAppCompatible: AppiumCapability {
    // Provided methods
    fn bundle_id(&mut self, id: &str) { ... }
    fn localizable_strings_dir(&mut self, dir: &str) { ... }
    fn language(&mut self, language: &str) { ... }
    fn locale(&mut self, locale: &str) { ... }
    fn calendar_format(&mut self, value: &str) { ... }
    fn app_push_timeout(&mut self, duration: Duration) { ... }
    fn app_install_strategy(&mut self, value: &str) { ... }
    fn auto_accept_alerts(&mut self, value: bool) { ... }
}

Capabilities for XCUITest (iOS).

Provided Methods

fn bundle_id(&mut self, id: &str)

Bundle id of app. Looks like app package (com.my.app).

fn localizable_strings_dir(&mut self, dir: &str)

Where to look for localizable strings. Default en.lproj

fn language(&mut self, language: &str)

Language to set for the simulator / emulator.

You need to set this manually on physical devices.

fn locale(&mut self, locale: &str)

Locale to set for the simulator / emulator.

You need to set this manually on physical devices.

fn calendar_format(&mut self, value: &str)

Calendar format to set for the iOS Simulator (eg. gregorian).

fn app_push_timeout(&mut self, duration: Duration)

Timeout for application upload in millisecond, on real devices

fn app_install_strategy(&mut self, value: &str)

Select application installation strategy for real devices.

The following strategies are supported:

serial (default) - pushes app files to the device in a sequential order; this is the least performant strategy, although the most reliable;

parallel - pushes app files simultaneously; this is usually the the most performant strategy, but sometimes could not be very stable;

ios-deploy - tells the driver to use a third-party tool ios-deploy to install the app; obviously the tool must be installed separately first and must be present in PATH before it could be used.

fn auto_accept_alerts(&mut self, value: bool)

Accept all iOS alerts automatically if they pop up.

This includes privacy access permission alerts (e.g., location, contacts, photos). Default is false.