Enum appium_client::commands::keyboard::AndroidKey

#[repr(u16)]
pub enum AndroidKey {

    Unknown = 0,
    SoftLeft = 1,
    SoftRight = 2,
    Home = 3,
    Back = 4,
    Call = 5,
    EndCall = 6,
    Digit0 = 7,
    Digit1 = 8,
    Digit2 = 9,
    Digit3 = 10,
    Digit4 = 11,
    Digit5 = 12,
    Digit6 = 13,
    Digit7 = 14,
    Digit8 = 15,
    Digit9 = 16,
    Asterisk = 17,
    Pound = 18,
    DPadUp = 19,
    DPadDown = 20,
    DPadLeft = 21,
    DpadRight = 22,
    DPadCenter = 23,
    VolumeUp = 24,
    VolumeDown = 25,
    Power = 26,
    Camera = 27,
    Clear = 28,
    A = 29,
    B = 30,
    C = 31,
    D = 32,
    E = 33,
    F = 34,
    G = 35,
    H = 36,
    I = 37,
    J = 38,
    K = 39,
    L = 40,
    M = 41,
    N = 42,
    O = 43,
    P = 44,
    Q = 45,
    R = 46,
    S = 47,
    T = 48,
    U = 49,
    V = 50,
    W = 51,
    X = 52,
    Y = 53,
    Z = 54,
    Comma = 55,
    Period = 56,
    AltLeft = 57,
    AltRight = 58,
    ShiftLeft = 59,
    ShiftRight = 60,
    Tab = 61,
    Space = 62,
    SYM = 63,
    Explorer = 64,
    Envelope = 65,
    Enter = 66,
    DEL = 67,
    Grave = 68,
    Minus = 69,
    Equals = 70,
    LeftBracket = 71,
    RightBracket = 72,
    Backslash = 73,
    Semicolon = 74,
    Apostrophe = 75,
    Slash = 76,
    At = 77,
    NUM = 78,
    HeadsetHook = 79,
    Focus = 80,
    Plus = 81,
    Menu = 82,
    Notification = 83,
    Search = 84,
    MediaPlayPause = 85,
    MediaStop = 86,
    MediaNext = 87,
    MediaPrevious = 88,
    MediaRewind = 89,
    MediaFastForward = 90,
    Mute = 91,
    PageUp = 92,
    PageDown = 93,
    PictSymbols = 94,
    SwitchCharset = 95,
    ButtonA = 96,
    ButtonB = 97,
    ButtonC = 98,
    ButtonX = 99,
    ButtonY = 100,
    ButtonZ = 101,
    ButtonL1 = 102,
    ButtonR1 = 103,
    ButtonL2 = 104,
    ButtonR2 = 105,
    ButtonThumbL = 106,
    ButtonThumbR = 107,
    ButtonStart = 108,
    ButtonSelect = 109,
    ButtonMode = 110,
    Escape = 111,
    ForwardDel = 112,
    CtrlLeft = 113,
    CtrlRight = 114,
    CapsLock = 115,
    ScrollLock = 116,
    MetaLeft = 117,
    MetaRight = 118,
    Function = 119,
    SYSRQ = 120,
    Break = 121,
    MoveHome = 122,
    MoveEnd = 123,
    Insert = 124,
    Forward = 125,
    MediaPlay = 126,
    MediaPause = 127,
    MediaClose = 128,
    MediaEject = 129,
    MediaRecord = 130,
    F1 = 131,
    F2 = 132,
    F3 = 133,
    F4 = 134,
    F5 = 135,
    F6 = 136,
    F7 = 137,
    F8 = 138,
    F9 = 139,
    F10 = 140,
    F11 = 141,
    F12 = 142,
    NumLock = 143,
    Numpad0 = 144,
    Numpad1 = 145,
    Numpad2 = 146,
    Numpad3 = 147,
    Numpad4 = 148,
    Numpad5 = 149,
    Numpad6 = 150,
    Numpad7 = 151,
    Numpad8 = 152,
    Numpad9 = 153,
    NumpadDivide = 154,
    NumpadMultiply = 155,
    NumpadSubtract = 156,
    NumpadAdd = 157,
    NumpadDot = 158,
    NumpadComma = 159,
    NumpadEnter = 160,
    NumpadEquals = 161,
    NumpadLeftParen = 162,
    NumpadRightParen = 163,
    VolumeMute = 164,
    Info = 165,
    ChannelUp = 166,
    ChannelDown = 167,
    KeycodeZoomIn = 168,
    KeycodeZoomOut = 169,
    TV = 170,
    Window = 171,
    Guide = 172,
    DVR = 173,
    Bookmark = 174,
    Captions = 175,
    Settings = 176,
    TVPower = 177,
    TVInput = 178,
    STBPower = 179,
    STBInput = 180,
    AVRPower = 181,
    AVRInput = 182,
    ProgRed = 183,
    ProgGreen = 184,
    ProgYellow = 185,
    ProgBlue = 186,
    AppSwitch = 187,
    Button1 = 188,
    Button2 = 189,
    Button3 = 190,
    Button4 = 191,
    Button5 = 192,
    Button6 = 193,
    Button7 = 194,
    Button8 = 195,
    Button9 = 196,
    Button10 = 197,
    Button11 = 198,
    Button12 = 199,
    Button13 = 200,
    Button14 = 201,
    Button15 = 202,
    Button16 = 203,
    LanguageSwitch = 204,
    MannerMode = 205,
    Mode3D = 206,
    Contacts = 207,
    Calendar = 208,
    Music = 209,
    Calculator = 210,
    ZenkakuHankaku = 211,
    Eisu = 212,
    Muhenkan = 213,
    Henkan = 214,
    KatakanaHiragana = 215,
    Yen = 216,
    Ro = 217,
    Kana = 218,
    Assist = 219,
    BrightnessDown = 220,
    BrightnessUp = 221,
    MediaAudioTrack = 222,
    Sleep = 223,
    WakeUp = 224,
    Pairing = 225,
    MediaTopMenu = 226,
    Key11 = 227,
    Key12 = 228,
    LastChannel = 229,
    TVDataService = 230,
    VoiceAssist = 231,
    TVRadioService = 232,
    TVTeletext = 233,
    TVNumberEntry = 234,
    TVTerrestrialAnalog = 235,
    TVTerrestrialDigital = 236,
    TVSatellite = 237,
    TVSatelliteBS = 238,
    TVSatelliteCS = 239,
    TVSatelliteService = 240,
    TVNetwork = 241,
    TVAntennaCable = 242,
    TVInputHdmi1 = 243,
    TVInputHdmi2 = 244,
    TVInputHdmi3 = 245,
    TVInputHdmi4 = 246,
    TVInputComposite1 = 247,
    TVInputComposite2 = 248,
    TVInputComponent1 = 249,
    TVInputComponent2 = 250,
    TVInputVga1 = 251,
    TVAudioDescription = 252,
    TVAudioDescriptionMixUp = 253,
    TVAudioDescriptionMixDown = 254,
    TVZoomMode = 255,
    TVContentsMenu = 256,
    TVMediaContextMenu = 257,
    TVTimerProgramming = 258,
    Help = 259,
    NavigatePrevious = 260,
    NavigateNext = 261,
    NavigateIn = 262,
    NavigateOut = 263,
    StemPrimary = 264,
    Stem1 = 265,
    Stem2 = 266,
    Stem3 = 267,
    DpadUpLeft = 268,
    DpadDownLeft = 269,
    DpadUpRight = 270,
    DpadDownRight = 271,
    MediaSkipForward = 272,
    MediaSkipBackward = 273,
    MediaStepForward = 274,
    MediaStepBackward = 275,
    SoftSleep = 276,
    Cut = 277,
    Copy = 278,
    Paste = 279,

}

Variants

Unknown = 0

Key code constant: Unknown key code.

SoftLeft = 1

Key code constant: Soft Left key. Usually situated below the display on phones and used as a multi-function feature key for selecting a software defined function shown on the bottom left of the display.

SoftRight = 2

Key code constant: Soft Right key. Usually situated below the display on phones and used as a multi-function feature key for selecting a software defined function shown on the bottom right of the display.

Home = 3

Key code constant: Home key. This key is handled by the framework and is never delivered to applications.

Back = 4

Key code constant: Back key.

Call = 5

Key code constant: Call key.

EndCall = 6

Key code constant: End Call key.

Digit0 = 7

Key code constant: ‘0’ key.

Digit1 = 8

Key code constant: ‘1’ key.

Digit2 = 9

Key code constant: ‘2’ key.

Digit3 = 10

Key code constant: ‘3’ key.

Digit4 = 11

Key code constant: ‘4’ key.

Digit5 = 12

Key code constant: ‘5’ key.

Digit6 = 13

Key code constant: ‘6’ key.

Digit7 = 14

Key code constant: ‘7’ key.

Digit8 = 15

Key code constant: ‘8’ key.

Digit9 = 16

Key code constant: ‘9’ key.

Asterisk = 17

Key code constant: ‘*’ key.

Pound = 18

Key code constant: ‘#’ key.

DPadUp = 19

Key code constant: Directional Pad Up key. May also be synthesized from trackball motions.

DPadDown = 20

Key code constant: Directional Pad Down key. May also be synthesized from trackball motions.

DPadLeft = 21

Key code constant: Directional Pad Left key. May also be synthesized from trackball motions.

DpadRight = 22

Key code constant: Directional Pad Right key. May also be synthesized from trackball motions.

DPadCenter = 23

Key code constant: Directional Pad Center key. May also be synthesized from trackball motions.

VolumeUp = 24

Key code constant: Volume Up key. Adjusts the speaker volume up.

VolumeDown = 25

Key code constant: Volume Down key. Adjusts the speaker volume down.

Power = 26

Key code constant: Power key.

Camera = 27

Key code constant: Camera key. Used to launch a camera application or take pictures.

Clear = 28

Key code constant: Clear key.

A = 29

Key code constant: ‘A’ key.

B = 30

Key code constant: ‘B’ key.

C = 31

Key code constant: ‘C’ key.

D = 32

Key code constant: ‘D’ key.

E = 33

Key code constant: ‘E’ key.

F = 34

Key code constant: ‘F’ key.

G = 35

Key code constant: ‘G’ key.

H = 36

Key code constant: ‘H’ key.

I = 37

Key code constant: ‘I’ key.

J = 38

Key code constant: ‘J’ key.

K = 39

Key code constant: ‘K’ key.

L = 40

Key code constant: ‘L’ key.

M = 41

Key code constant: ‘M’ key.

N = 42

Key code constant: ‘N’ key.

O = 43

Key code constant: ‘O’ key.

P = 44

Key code constant: ‘P’ key.

Q = 45

Key code constant: ‘Q’ key.

R = 46

Key code constant: ‘R’ key.

S = 47

Key code constant: ‘S’ key.

T = 48

Key code constant: ‘T’ key.

U = 49

Key code constant: ‘U’ key.

V = 50

Key code constant: ‘V’ key.

W = 51

Key code constant: ‘W’ key.

X = 52

Key code constant: ‘X’ key.

Y = 53

Key code constant: ‘Y’ key.

Z = 54

Key code constant: ‘Z’ key.

Comma = 55

Key code constant: ‘,’ key.

Period = 56

Key code constant: ‘.’ key.

AltLeft = 57

Key code constant: Left Alt modifier key.

AltRight = 58

Key code constant: Right Alt modifier key.

ShiftLeft = 59

Key code constant: Left Shift modifier key.

ShiftRight = 60

Key code constant: Right Shift modifier key.

Tab = 61

Key code constant: Tab key.

Space = 62

Key code constant: Space key.

SYM = 63

Key code constant: Symbol modifier key. Used to enter alternate symbols.

Explorer = 64

Key code constant: Explorer special function key. Used to launch a browser application.

Envelope = 65

Key code constant: Envelope special function key. Used to launch a mail application.

Enter = 66

Key code constant: Enter key.

DEL = 67

Key code constant: Backspace key. Deletes characters before the insertion point, unlike AndroidKey::ForwardDel.

Grave = 68

Key code constant: ‘`’ (backtick) key.

Minus = 69

Key code constant: ‘-’.

Equals = 70

Key code constant: ‘=’ key.

LeftBracket = 71

Key code constant: ‘[’ key.

RightBracket = 72

Key code constant: ‘]’ key.

Backslash = 73

Key code constant: ’' key.

Semicolon = 74

Key code constant: ‘;’ key.

Apostrophe = 75

Key code constant: ‘’’ (apostrophe) key.

Slash = 76

Key code constant: ‘/’ key.

At = 77

Key code constant: ‘@’ key.

NUM = 78

Key code constant: Number modifier key. Used to enter numeric symbols. This key is not Num Lock; it is more like AndroidKey::AltLeft and is interpreted as an ALT key

HeadsetHook = 79

Key code constant: Headset Hook key. Used to hang up calls and stop media.

Focus = 80

Key code constant: Camera Focus key. Used to focus the camera.

Plus = 81

Key code constant: ‘+’ key.

Menu = 82

Key code constant: Menu key.

Notification = 83

Key code constant: Notification key.

Search = 84

Key code constant: Search key.

MediaPlayPause = 85

Key code constant: Play/Pause media key.

MediaStop = 86

Key code constant: Stop media key.

MediaNext = 87

Key code constant: Play Next media key.

MediaPrevious = 88

Key code constant: Play Previous media key.

MediaRewind = 89

Key code constant: Rewind media key.

MediaFastForward = 90

Key code constant: Fast Forward media key.

Mute = 91

Key code constant: Mute key. Mutes the microphone, unlike AndroidKey::VolumeMute.

PageUp = 92

Key code constant: Page Up key.

PageDown = 93

Key code constant: Page Down key.

PictSymbols = 94

Key code constant: Picture Symbols modifier key. Used to switch symbol sets (Emoji, Kao-moji).

SwitchCharset = 95

Key code constant: Switch Charset modifier key. Used to switch character sets (Kanji, Katakana).

ButtonA = 96

Key code constant: A Button key. On a game controller, the A button should be either the button labeled A or the first button on the bottom row of controller buttons.

ButtonB = 97

Key code constant: B Button key. On a game controller, the B button should be either the button labeled B or the second button on the bottom row of controller buttons.

ButtonC = 98

Key code constant: C Button key. On a game controller, the C button should be either the button labeled C or the third button on the bottom row of controller buttons.

ButtonX = 99

Key code constant: X Button key. On a game controller, the X button should be either the button labeled X or the first button on the upper row of controller buttons.

ButtonY = 100

Key code constant: Y Button key. On a game controller, the Y button should be either the button labeled Y or the second button on the upper row of controller buttons.

ButtonZ = 101

Key code constant: Z Button key. On a game controller, the Z button should be either the button labeled Z or the third button on the upper row of controller buttons.

ButtonL1 = 102

Key code constant: L1 Button key. On a game controller, the L1 button should be either the button labeled L1 (or L) or the top left trigger button.

ButtonR1 = 103

Key code constant: R1 Button key. On a game controller, the R1 button should be either the button labeled R1 (or R) or the top right trigger button.

ButtonL2 = 104

Key code constant: L2 Button key. On a game controller, the L2 button should be either the button labeled L2 or the bottom left trigger button.

ButtonR2 = 105

Key code constant: R2 Button key. On a game controller, the R2 button should be either the button labeled R2 or the bottom right trigger button.

ButtonThumbL = 106

Key code constant: Left Thumb Button key. On a game controller, the left thumb button indicates that the left (or only) joystick is pressed.

ButtonThumbR = 107

Key code constant: Right Thumb Button key. On a game controller, the right thumb button indicates that the right joystick is pressed.

ButtonStart = 108

Key code constant: Start Button key. On a game controller, the button labeled Start.

ButtonSelect = 109

Key code constant: Select Button key. On a game controller, the button labeled Select.

ButtonMode = 110

Key code constant: Mode Button key. On a game controller, the button labeled Mode.

Escape = 111

Key code constant: Escape key.

ForwardDel = 112

Key code constant: Forward Delete key. Deletes characters ahead of the insertion point, unlike AndroidKey::DEL.

CtrlLeft = 113

Key code constant: Left Control modifier key.

CtrlRight = 114

Key code constant: Right Control modifier key.

CapsLock = 115

Key code constant: Caps Lock key.

ScrollLock = 116

Key code constant: Scroll Lock key.

MetaLeft = 117

Key code constant: Left Meta modifier key.

MetaRight = 118

Key code constant: Right Meta modifier key.

Function = 119

Key code constant: Function modifier key.

SYSRQ = 120

Key code constant: System Request / Print Screen key.

Break = 121

Key code constant: Break / Pause key.

MoveHome = 122

Key code constant: Home Movement key. Used for scrolling or moving the cursor around to the start of a line or to the top of a list.

MoveEnd = 123

Key code constant: End Movement key. Used for scrolling or moving the cursor around to the end of a line or to the bottom of a list.

Insert = 124

Key code constant: Insert key. Toggles insert / overwrite edit mode.

Forward = 125

Key code constant: Forward key. Navigates forward in the history stack. Complement of AndroidKey::Back.

MediaPlay = 126

Key code constant: Play media key.

MediaPause = 127

Key code constant: Pause media key.

MediaClose = 128

Key code constant: Close media key. May be used to close a CD tray, for example.

MediaEject = 129

Key code constant: Eject media key. May be used to eject a CD tray, for example.

MediaRecord = 130

Key code constant: Record media key.

F1 = 131

Key code constant: F1 key.

F2 = 132

Key code constant: F2 key.

F3 = 133

Key code constant: F3 key.

F4 = 134

Key code constant: F4 key.

F5 = 135

Key code constant: F5 key.

F6 = 136

Key code constant: F6 key.

F7 = 137

Key code constant: F7 key.

F8 = 138

Key code constant: F8 key.

F9 = 139

Key code constant: F9 key.

F10 = 140

Key code constant: F10 key.

F11 = 141

Key code constant: F11 key.

F12 = 142

Key code constant: F12 key.

NumLock = 143

Key code constant: Num Lock key. This is the Num Lock key; it is different from AndroidKey::NUM. This key alters the behavior of other keys on the numeric keypad.

Numpad0 = 144

Key code constant: Numeric keypad ‘0’ key.

Numpad1 = 145

Key code constant: Numeric keypad ‘1’ key.

Numpad2 = 146

Key code constant: Numeric keypad ‘2’ key.

Numpad3 = 147

Key code constant: Numeric keypad ‘3’ key.

Numpad4 = 148

Key code constant: Numeric keypad ‘4’ key.

Numpad5 = 149

Key code constant: Numeric keypad ‘5’ key.

Numpad6 = 150

Key code constant: Numeric keypad ‘6’ key.

Numpad7 = 151

Key code constant: Numeric keypad ‘7’ key.

Numpad8 = 152

Key code constant: Numeric keypad ‘8’ key.

Numpad9 = 153

Key code constant: Numeric keypad ‘9’ key.

NumpadDivide = 154

Key code constant: Numeric keypad ‘/’ key (for division).

NumpadMultiply = 155

Key code constant: Numeric keypad ‘*’ key (for multiplication).

NumpadSubtract = 156

Key code constant: Numeric keypad ‘-’ key (for subtraction).

NumpadAdd = 157

Key code constant: Numeric keypad ‘+’ key (for addition).

NumpadDot = 158

Key code constant: Numeric keypad ‘.’ key (for decimals or digit grouping).

NumpadComma = 159

Key code constant: Numeric keypad ‘,’ key (for decimals or digit grouping).

NumpadEnter = 160

Key code constant: Numeric keypad Enter key.

NumpadEquals = 161

Key code constant: Numeric keypad ‘=’ key.

NumpadLeftParen = 162

Key code constant: Numeric keypad ‘(’ key.

NumpadRightParen = 163

Key code constant: Numeric keypad ‘)’ key.

VolumeMute = 164

Key code constant: Volume Mute key. Mutes the speaker, unlike AndroidKey::Mute. This key should normally be implemented as a toggle such that the first press mutes the speaker and the second press restores the original volume.

Info = 165

Key code constant: Info key. Common on TV remotes to show additional information related to what is currently being viewed.

ChannelUp = 166

Key code constant: Channel up key. On TV remotes, increments the television channel.

ChannelDown = 167

Key code constant: Channel down key. On TV remotes, decrements the television channel.

KeycodeZoomIn = 168

Key code constant: Zoom in key.

KeycodeZoomOut = 169

Key code constant: Zoom out key.

TV = 170

Key code constant: TV key. On TV remotes, switches to viewing live TV.

Window = 171

Key code constant: Window key. On TV remotes, toggles picture-in-picture mode or other windowing functions.

Guide = 172

Key code constant: Guide key. On TV remotes, shows a programming guide.

DVR = 173

Key code constant: DVR key. On some TV remotes, switches to a DVR mode for recorded shows.

Bookmark = 174

Key code constant: Bookmark key. On some TV remotes, bookmarks content or web pages.

Captions = 175

Key code constant: Toggle captions key. Switches the mode for closed-captioning text, for example during television shows.

Settings = 176

Key code constant: Settings key. Starts the system settings activity.

TVPower = 177

Key code constant: TV power key. On TV remotes, toggles the power on a television screen.

TVInput = 178

Key code constant: TV input key. On TV remotes, switches the input on a television screen.

STBPower = 179

Key code constant: Set-top-box power key. On TV remotes, toggles the power on an external Set-top-box.

STBInput = 180

Key code constant: Set-top-box input key. On TV remotes, switches the input mode on an external Set-top-box.

AVRPower = 181

Key code constant: A/V Receiver power key. On TV remotes, toggles the power on an external A/V Receiver.

AVRInput = 182

Key code constant: A/V Receiver input key. On TV remotes, switches the input mode on an external A/V Receiver.

ProgRed = 183

Key code constant: Red “programmable” key. On TV remotes, acts as a contextual/programmable key.

ProgGreen = 184

Key code constant: Green “programmable” key. On TV remotes, actsas a contextual/programmable key.

ProgYellow = 185

Key code constant: Yellow “programmable” key. On TV remotes, acts as a contextual/programmable key.

ProgBlue = 186

Key code constant: Blue “programmable” key. On TV remotes, acts as a contextual/programmable key.

AppSwitch = 187

Key code constant: App switch key. Should bring up the application switcher dialog.

Button1 = 188

Key code constant: Generic Game Pad Button #1.

Button2 = 189

Key code constant: Generic Game Pad Button #2.

Button3 = 190

Key code constant: Generic Game Pad Button #3.

Button4 = 191

Key code constant: Generic Game Pad Button #4.

Button5 = 192

Key code constant: Generic Game Pad Button #5.

Button6 = 193

Key code constant: Generic Game Pad Button #6.

Button7 = 194

Key code constant: Generic Game Pad Button #7.

Button8 = 195

Key code constant: Generic Game Pad Button #8.

Button9 = 196

Key code constant: Generic Game Pad Button #9.

Button10 = 197

Key code constant: Generic Game Pad Button #10.

Button11 = 198

Key code constant: Generic Game Pad Button #11.

Button12 = 199

Key code constant: Generic Game Pad Button #12.

Button13 = 200

Key code constant: Generic Game Pad Button #13.

Button14 = 201

Key code constant: Generic Game Pad Button #14.

Button15 = 202

Key code constant: Generic Game Pad Button #15.

Button16 = 203

Key code constant: Generic Game Pad Button #16.

LanguageSwitch = 204

Key code constant: Language Switch key. Toggles the current input language such as switching between English and Japanese on a QWERTY keyboard. On some devices, the same function may be performed by pressing Shift+Spacebar.

MannerMode = 205

Key code constant: Manner Mode key. Toggles silent or vibrate mode on and off to make the device behave more politely in certain settings such as on a crowded train. On some devices, the key may only operate when long-pressed.

Mode3D = 206

Key code constant: 3D Mode key. Toggles the display between 2D and 3D mode.

Contacts = 207

Key code constant: Contacts special function key. Used to launch an address book application.

Calendar = 208

Key code constant: Calendar special function key. Used to launch a calendar application.

Music = 209

Key code constant: Music special function key. Used to launch a music player application.

Calculator = 210

Key code constant: Calculator special function key. Used to launch a calculator application.

ZenkakuHankaku = 211

Key code constant: Japanese full-width / half-width key.

Eisu = 212

Key code constant: Japanese alphanumeric key.

Muhenkan = 213

Key code constant: Japanese non-conversion key.

Henkan = 214

Key code constant: Japanese conversion key.

KatakanaHiragana = 215

Key code constant: Japanese katakana / hiragana key.

Yen = 216

Key code constant: Japanese Yen key.

Ro = 217

Key code constant: Japanese Ro key.

Kana = 218

Key code constant: Japanese kana key.

Assist = 219

Key code constant: Assist key. Launches the global assist activity. Not delivered to applications.

BrightnessDown = 220

Key code constant: Brightness Down key. Adjusts the screen brightness down.

BrightnessUp = 221

Key code constant: Brightness Up key. Adjusts the screen brightness up.

MediaAudioTrack = 222

Key code constant: Audio Track key. Switches the audio tracks.

Sleep = 223

Key code constant: Sleep key. Puts the device to sleep. Behaves somewhat like AndroidKey::Power but it has no effect if the device is already asleep.

WakeUp = 224

Key code constant: Wakeup key. Wakes up the device. Behaves somewhat like AndroidKey::Power but it has no effect if the device is already awake.

Pairing = 225

Key code constant: Pairing key. Initiates peripheral pairing mode. Useful for pairing remote control devices or game controllers, especially if no other input mode is available.

MediaTopMenu = 226

Key code constant: Media Top Menu key. Goes to the top of media menu.

Key11 = 227

Key code constant: ‘11’ key.

Key12 = 228

Key code constant: ‘12’ key.

LastChannel = 229

Key code constant: Last Channel key. Goes to the last viewed channel.

TVDataService = 230

Key code constant: TV data service key. Displays data services like weather, sports.

VoiceAssist = 231

Key code constant: Voice Assist key. Launches the global voice assist activity. Not delivered to applications.

TVRadioService = 232

Key code constant: Radio key. Toggles TV service / Radio service.

TVTeletext = 233

Key code constant: Teletext key. Displays Teletext service.

TVNumberEntry = 234

Key code constant: Number entry key. Initiates to enter multi-digit channel nubmber when each digit key is assigned for selecting separate channel. Corresponds to Number Entry Mode (0x1D) of CEC User Control Code.

TVTerrestrialAnalog = 235

Key code constant: Analog Terrestrial key. Switches to analog terrestrial broadcast service.

TVTerrestrialDigital = 236

Key code constant: Digital Terrestrial key. Switches to digital terrestrial broadcast service.

TVSatellite = 237

Key code constant: Satellite key. Switches to digital satellite broadcast service.

TVSatelliteBS = 238

Key code constant: BS key. Switches to BS digital satellite broadcasting service available in Japan.

TVSatelliteCS = 239

Key code constant: CS key. Switches to CS digital satellite broadcasting service available in Japan.

TVSatelliteService = 240

Key code constant: BS/CS key. Toggles between BS and CS digital satellite services.

TVNetwork = 241

Key code constant: Toggle Network key. Toggles selecting broacast services.

TVAntennaCable = 242

Key code constant: Antenna/Cable key. Toggles broadcast input source between antenna and cable.

TVInputHdmi1 = 243

Key code constant: HDMI #1 key. Switches to HDMI input #1.

TVInputHdmi2 = 244

Key code constant: HDMI #2 key. Switches to HDMI input #2.

TVInputHdmi3 = 245

Key code constant: HDMI #3 key. Switches to HDMI input #3.

TVInputHdmi4 = 246

Key code constant: HDMI #4 key. Switches to HDMI input #4.

TVInputComposite1 = 247

Key code constant: Composite #1 key. Switches to composite video input #1.

TVInputComposite2 = 248

Key code constant: Composite #2 key. Switches to composite video input #2.

TVInputComponent1 = 249

Key code constant: Component #1 key. Switches to component video input #1.

TVInputComponent2 = 250

Key code constant: Component #2 key. Switches to component video input #2.

TVInputVga1 = 251

Key code constant: VGA #1 key. Switches to VGA (analog RGB) input #1.

TVAudioDescription = 252

Key code constant: Audio description key. Toggles audio description off / on.

TVAudioDescriptionMixUp = 253

Key code constant: Audio description mixing volume up key. Louden audio description volume as compared with normal audio volume.

TVAudioDescriptionMixDown = 254

Key code constant: Audio description mixing volume down key. Lessen audio description volume as compared with normal audio volume.

TVZoomMode = 255

Key code constant: Zoom mode key. Changes Zoom mode (Normal, Full, Zoom, Wide-zoom, etc.)

TVContentsMenu = 256

Key code constant: Contents menu key. Goes to the title list. Corresponds to Contents Menu (0x0B) of CEC User Control Code

TVMediaContextMenu = 257

Key code constant: Media context menu key. Goes to the context menu of media contents. Corresponds to Media Context-sensitive Menu (0x11) of CEC User Control Code.

TVTimerProgramming = 258

Key code constant: Timer programming key. Goes to the timer recording menu. Corresponds to Timer Programming (0x54) of CEC User Control Code.

Help = 259

Key code constant: Help key.

NavigatePrevious = 260

Key code constant: Navigate to previous key. Goes backward by one item in an ordered collection of items.

NavigateNext = 261

Key code constant: Navigate to next key. Advances to the next item in an ordered collection of items.

NavigateIn = 262

Key code constant: Navigate in key. Activates the item that currently has focus or expands to the next level of a navigation hierarchy.

NavigateOut = 263

Key code constant: Navigate out key. Backs out one level of a navigation hierarchy or collapses the item that currently has focus.

StemPrimary = 264

Key code constant: Primary stem key for Wear. Main power/reset button on watch.

Stem1 = 265

Key code constant: Generic stem key 1 for Wear.

Stem2 = 266

Key code constant: Generic stem key 2 for Wear.

Stem3 = 267

Key code constant: Generic stem key 3 for Wear.

DpadUpLeft = 268

Key code constant: Directional Pad Up-Left.

DpadDownLeft = 269

Key code constant: Directional Pad Down-Left.

DpadUpRight = 270

Key code constant: Directional Pad Up-Right.

DpadDownRight = 271

Key code constant: Directional Pad Down-Right.

MediaSkipForward = 272

Key code constant: Skip forward media key.

MediaSkipBackward = 273

Key code constant: Skip backward media key.

MediaStepForward = 274

Key code constant: Step forward media key. Steps media forward, one frame at a time.

MediaStepBackward = 275

Key code constant: Step backward media key. Steps media backward, one frame at a time.

SoftSleep = 276

Key code constant: put device to sleep unless a wakelock is held.

Cut = 277

Key code constant: Cut key.

Copy = 278

Key code constant: Copy key.

Paste = 279

Key code constant: Paste key.