Struct appium_client::Client

pub struct Client<Caps>where
    Caps: AppiumCapability,{ /* private fields */ }

Generic Appium client

This client represents an Appium client that will connect to an Appium server and send command to said server.

Depending on chosen capabilities (AppiumCapability), the client will have different traits. Which means - different available features.

Check out AndroidClient and IOSClient in docs to see their features (available commands).

Note: Client automatically ends Appium session on drop (end of lifetime). This is the only way to end session.