Type Alias appium_client::AndroidClient

pub type AndroidClient = Client<AndroidCapabilities>;

Client used to automate Android testing

To create AndroidClient, you need to use ClientBuilder and AndroidCapabilities. Rust type system will automatically pick up that by using those capabilities, you mean to control an Android device.

See trait implementations to check available features (commands) of this client.

use appium_client::capabilities::{AppCapable, UdidCapable, UiAutomator2AppCompatible};
use appium_client::capabilities::android::AndroidCapabilities;
use appium_client::ClientBuilder;

let mut capabilities = AndroidCapabilities::new_uiautomator();
capabilities.udid("emulator-5554");
capabilities.app("/apps/sample.apk");
capabilities.app_wait_activity("com.example.AppActivity");

let client = ClientBuilder::native(capabilities)
   .connect("http://localhost:4723/wd/hub/")
   .await?;

// congratulations, you have successfully created an AndroidClient

Aliased Type

struct AndroidClient { /* private fields */ }