Type Alias appium_client::IOSClient

pub type IOSClient = Client<IOSCapabilities>;

Client used to automate iOS testing

To create IOSClient, you need to use ClientBuilder and IOSCapabilities. Rust type system will automatically pick up that by using those capabilities, you mean to control an iOS device.

See trait implementations to check available features (commands) of this client.

use appium_client::capabilities::{AppCapable, UdidCapable};
use appium_client::capabilities::ios::IOSCapabilities;
use appium_client::ClientBuilder;

let mut capabilities = IOSCapabilities::new_xcui();
capabilities.udid("000011114567899");
capabilities.app("/apps/sample.ipa");

let client = ClientBuilder::native(capabilities)
   .connect("http://localhost:4723/wd/hub/")
   .await?;

// congratulations, you have successfully created an IOSClient

Aliased Type

struct IOSClient { /* private fields */ }